go/staking/api: Extract token-related things to go/staking/api/token
So it can be imported by other packages without causing import cycles.

go/staking/api: Move token-related context keys to
go/common/prettyprint.
