+++ .changelog/3167.feature.1.md
@@ -1,0 +1,1 @@
+go/staking/api/token: Initial implementation
